requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.16
langchain-core>=0.1.18
langchain-openai>=0.0.5
langchain-text-splitters>=0.0.1
langchain-pinecone>=0.0.2
langchain-postgres>=0.0.1
faiss-cpu>=1.7.4
pinecone-client>=3.0.0
psycopg2-binary>=2.9.9
ftfy>=6.1.3
python-multipart>=0.0.6
beautifulsoup4>=4.12.0
docx2txt>=0.8
pypdf>=4.0.0
openpyxl>=3.1.0
python-pptx>=0.6.0
unstructured>=0.11.0
markdown>=3.5.0
docstring-parser>=0.15
tqdm>=4.66.0